requirements.txt
farm-haystack
gradio
streamlit
fastapi
uvicorn 
python-multipart 
jinja2